requirements.txt
google-generativeai
python-dotenv
pgvector
sqlalchemy
psycopg2
pandas